text-to-sql/google_spreadsheet_to_md/pyproject.toml
[project]
name = "sheets-to-csv"
version = "0.1.0"
description = "Convert Google Sheets to CSV"
authors = [
    {name = "ckifonidis@gmal.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-auth==2.27.0",
    "google-auth-oauthlib==1.2.0",
    "google-auth-httplib2==0.2.0",
    "google-api-python-client==2.116.0",
    "pandas==2.2.0",
    "python-dotenv==1.0.1"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
